\name{cos_similarity}
\alias{cos_similarity}
\title{Cosine similarity between peaksets}
\description{
  This function calculate cos similarity between two sets of peak ranges
}
\usage{
cos_similarity(x, y)
}
\arguments{
  \item{x, y}{\code{GRanges} instances representing peak ranges }
}
\value{
  The cosine similarity of \code{x} and \code{y}.
}
\examples{

}
